Move utilty functions into their own module
